+++ apps/wiki/app/(routes)/merkevare/ikoner/ikonsett/page.tsx
@@ -1,5 +1,6 @@
 import { PageDetails } from "@/types/types"
 import { pagePathname, placeholderImageUrl } from "@/lib/utils"
+import { IconGrid } from "./icon-grid"
 
 export const metadata: PageDetails = {
   href: pagePathname(import.meta.url),
@@ -7,9 +8,12 @@ export const metadata: PageDetails = {
   image: placeholderImageUrl({}),
   description:
     "Ikonsett er en viktig del av designsystemet vårt, og brukes for å visualisere informasjon og veilede brukeren. Ikoner skal være enkle, tydelige og konsistente, og skal brukes i tråd med retningslinjene våre.",
-  children: [],
 }
 
 export default function IkonSett() {
-  return <div>IkonSett</div>
+  return (
+    <div>
+      <IconGrid />
+    </div>
+  )
 }