+++ apps/wiki/app/(routes)/merkevare/ikoner/retningslinjer/page.tsx
@@ -7,7 +7,6 @@ export const metadata: PageDetails = {
   image: placeholderImageUrl({}),
   description:
     "Retningslinjer er en viktig del av designsystemet vårt, og brukes for å visualisere informasjon og veilede brukeren. Ikoner skal være enkle, tydelige og konsistente, og skal brukes i tråd med retningslinjene våre.",
-  children: [],
 }
 
 export default function RetningsLinjerPage() {