+++ apps/wiki/app/layout.tsx
@@ -4,7 +4,7 @@ import { SessionProvider } from "next-auth/react"
 import { APP_NAME } from "@/lib/constants"
 import { cn } from "@/lib/utils"
 import Footer from "@/components/footer"
-import { IconLoader, ThemeContext } from "@/components/kobber-ssr-loader"
+import { IconLoader } from "@/components/kobber-icons-loader"
 import { WikiNavbar } from "@/components/menu/wiki-navbar"
 
 /** Fallback for all pages */
@@ -23,17 +23,15 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <html lang="en" suppressHydrationWarning>
+    <html lang="no" suppressHydrationWarning>
       <body className={cn("bg-[#FDF9F9] text-[#481125ff] antialiased transition-all")}>
         <SessionProvider>
-          <ThemeContext themeId="kobber-theme-default">
-            <IconLoader />
-            <div className="mx-auto flex min-h-screen w-full max-w-max-width flex-col gap-y-page/gap/horizontal px-page/padding/inline/xsmall md:px-page/padding/inline/large">
-              <WikiNavbar />
-              {children}
-              <Footer />
-            </div>
-          </ThemeContext>
+          <IconLoader />
+          <div className="sm:px-page/padding/inline/small md:px-page/padding/inline/medium mx-auto flex min-h-screen w-full max-w-max-width flex-col gap-y-page/gap/horizontal px-page/padding/inline/xsmall xl:px-page/padding/inline/large">
+            <WikiNavbar />
+            {children}
+            <Footer />
+          </div>
         </SessionProvider>
       </body>
     </html>