+++ apps/wiki/components/kobber-components.tsx
@@ -1,32 +1,28 @@
 "use client"
 
-import { KobberArticleLayout, KobberBody, KobberIngress } from "@gyldendal/kobber-components/react"
+import {
+  KobberAccordion,
+  KobberButton,
+  KobberCardLayout,
+  KobberCardLayoutColumnAspectRatio,
+  KobberDivider,
+  KobberGrid,
+  KobberList,
+  KobberListItem,
+    KobberArticleLayout, KobberBody, KobberIngress
+} from "@gyldendal/kobber-components/react"
 
-export { KobberArticleLayout as ArticleLayout, KobberBody as Body, KobberIngress as Ingress }
-
-// import {
-//   KobberAccordion,
-//   KobberButton,
-//   KobberCardLayout,
-//   KobberCardLayoutColumnAspectRatio,
-//   KobberDivider,
-//   KobberGrid,
-//   KobberList,
-//   KobberListItem,
-//   KobberThemeContext,
-// } from "@gyldendal/kobber-components/react"
-// import { IconArrowRight } from "@gyldendal/kobber-icons/react"
-
-// export {
-//   KobberButton as Button,
-//   KobberCardLayout as CardLayout,
-//   KobberCardLayoutColumnAspectRatio as CardLayoutToColumnAspectRatio,
-//   KobberThemeContext as ThemeContext,
-//   KobberDivider as Divider,
-//   KobberAccordion as Accordion,
-//   KobberList as List,
-//   KobberListItem as ListItem,
-//   KobberGrid as Grid,
-// }
-
-// export { IconArrowRight }
+/**
+ * Wraps all kobber components in a "use client" file to prevent SSR errors.
+ */
+export {
+  KobberButton as Button,
+  KobberCardLayout as CardLayout,
+  KobberCardLayoutColumnAspectRatio as CardLayoutColumnAspectRatio,
+  KobberDivider as Divider,
+  KobberAccordion as Accordion,
+  KobberList as List,
+  KobberListItem as ListItem,
+  KobberGrid as Grid,
+    KobberArticleLayout as ArticleLayout, KobberBody as Body, KobberIngress as Ingress 
+}