+++ apps/wiki/components/menu/side-menu-item.tsx
@@ -2,7 +2,7 @@
 
 import { usePathname } from "next/navigation"
 import { PageDetails } from "@/types/types"
-import { List } from "@/components/kobber-ssr-loader"
+import { List } from "@/components/kobber-components"
 import { SideMenuItemLink } from "./side-menu-item-link"
 
 export const SideMenuItem = (item: PageDetails) => {
@@ -11,7 +11,7 @@ export const SideMenuItem = (item: PageDetails) => {
     <>
       <SideMenuItemLink {...item} />
 
-      {pathname?.includes(item.href) && item.children && (
+      {pathname?.includes(item.href) && item.children && item.children.length !== 0 && (
         <List orientation="vertical" className="pl-[20px]">
           {item.children.map((item) => (
             <SideMenuItem key={item.href} {...item} />