+++ apps/wiki/components/menu/side-menu.tsx
@@ -2,7 +2,6 @@
 
 import { Sidebar } from "lucide-react"
 import { useIsMobile } from "@/hooks/use-is-mobile"
-import { ThemeContext } from "@/components/kobber-ssr-loader"
 import { Button } from "../ui/button"
 import { Sheet, SheetContent, SheetTrigger } from "../ui/sheet"
 
@@ -13,8 +12,9 @@ type SideMenuProps = {
 export const SideMenu = ({ children }: SideMenuProps) => {
   const isMobile = useIsMobile()
 
+  // top-[96px] is the height of the navbar + gap
   return (
-    <aside className="flex h-fit flex-col gap-y-[8px] border-r border-[#E5CFD3] pb-1.5 pr-10">
+    <aside className="sticky top-[96px] flex flex-col gap-y-[8px] border-r border-[#E5CFD3] pb-1.5 pr-2">
       {isMobile ? (
         <Sheet modal={false}>
           <SheetTrigger asChild>
@@ -22,8 +22,8 @@ export const SideMenu = ({ children }: SideMenuProps) => {
               <Sidebar className="size-[16px]" />
             </Button>
           </SheetTrigger>
-          <SheetContent side={"left"} className="pt-72">
-            <ThemeContext themeId="kobber-theme-default">{children}</ThemeContext>
+          <SheetContent side={"left"} className="pt-[72px]">
+            {children}
           </SheetContent>
         </Sheet>
       ) : (