+++ apps/wiki/lib/auth.ts
@@ -3,7 +3,7 @@ import { accounts, users, verificationTokens } from "@/drizzle/schema"
 import MicrosoftEntraID from "@auth/core/providers/microsoft-entra-id"
 import { DrizzleAdapter } from "@auth/drizzle-adapter"
 import * as Sentry from "@sentry/nextjs"
-import NextAuth from "next-auth"
+import NextAuth, { AuthError } from "next-auth"
 
 const id = process.env.AUTH_MICROSOFT_ENTRA_ID_ISSUER
 
@@ -33,14 +33,19 @@ export const { auth, handlers, signIn, signOut } = NextAuth({
     }),
   ],
   logger: {
-    error(code, ...message) {
-      Sentry.captureException(new Error(`ERROR ${code}: ${message ? JSON.stringify(message) : ""}`))
+    error(error) {
+      const name = error instanceof AuthError ? error.type : error.name
+      Sentry.captureException(
+        new Error(
+          `ERROR AUTH ${name}: ${error.message}\n ${error.cause ? JSON.stringify(error.cause) : ""}\n ${error.stack ? JSON.stringify(error.stack) : ""}`
+        )
+      )
     },
-    warn(code, ...message) {
-      Sentry.captureMessage(`WARNING ${code}: ${message ? JSON.stringify(message) : ""}`)
-    },
-    debug(code, ...message) {
-      Sentry.captureMessage(`DEBUG ${code}: ${message ? JSON.stringify(message) : ""}`)
-    },
+    // warn(code) {
+    //   Sentry.captureMessage(`WARNING AUTH ${code}`, "warning")
+    // },
+    // debug(message, metadata) {
+    //   Sentry.captureMessage(`DEBUG AUTH ${message}: ${metadata ? JSON.stringify(metadata) : ""}`, "debug")
+    // },
   },
 })