+++ packages/kobber-base/build.ts
@@ -1,8 +1,11 @@
 import { themeConfigs } from "./buildConfig";
 import { buildTypography } from "./src/typography/buildTypography";
 import { buildThemeTokens } from "./src/styleDictionary/buildThemeTokens";
-import tokensFromFigma from "./tokens-from-figma";
+import { tokensFromFigma } from "./tokens-from-figma";
+
 /**
+ * Converts json from Figma into css, js, scss ++
+ *
  * Run by the build script in package.json
  */
 themeConfigs.forEach(themeConfig => {