+++ packages/kobber-components/src/divider/Divider.ts
@@ -1,20 +1,14 @@
-import { css, CSSResultGroup, html, LitElement } from "lit";
+import { CSSResultGroup, html } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import { themeContext } from "../utils/theme-context";
-import { consume } from "@lit/context";
-import { Theme } from "../utils/theme-context.types";
 import { DividerVariant } from "./Divider.types";
 import { dividerBaseStyles } from "./Divider.styles";
+import KobberElement from "../base/kobber-element";
 
 @customElement("kobber-divider")
-export class Divider extends LitElement {
-  @consume({ context: themeContext, subscribe: true })
-  theme?: Theme;
-
+export class Divider extends KobberElement {
   @property({ type: String })
   variant: DividerVariant = "main";
 
-  // Base styles
   static styles: CSSResultGroup = [dividerBaseStyles];
 
   render() {
@@ -23,20 +17,17 @@ export class Divider extends LitElement {
   }
 
   applyThemeStyles() {
-    const tokens = this.theme?.tokens;
-    if (!tokens) {
-      console.log("Should never occur");
-      return css``;
-    }
+    const component = this.tokens().component.divider;
 
     const style = document.createElement("style");
-    // TODO: Wait for tokens
+
     style.textContent = `
     :host {
-        --divider-main-background-color: #e5cfd3;
-        --divider-supplemental-background-color: #532d37;
+        --divider-main-background-color: ${component.background.color.main};
+        --divider-supplemental-background-color: ${component.background.color.supplemental};
     }
     `;
+
     this.shadowRoot?.appendChild(style);
   }
 }