+++ packages/kobber-components/src/index.web-components.ts
@@ -9,6 +9,12 @@ export { Checkbox } from "./checkbox/Checkbox";
 export { CardLayout } from "./layouts/card-layout/CardLayout";
 export { CardLayoutColumnAspectRatio } from "./layouts/card-layout/CardLayoutColumnAspectRatio";
 export { BoxLayout } from "./layouts/box-layout/BoxLayout";
+
 export { ArticleLayout } from "./article-layout/ArticleLayout";
 export { Body } from "./body/Body";
 export { Ingress } from "./ingress/Ingress";
+
+export { Carousel } from "./carousel/Carousel";
+export { HorizontalLayout } from "./layouts/horizontal-layout/HorizontalLayout";
+export { HorizontalLayoutColumn } from "./layouts/horizontal-layout/HorizontalLayoutColumn";
+